+++ src/components/LaborLecture/Desktop.js
@@ -6,18 +6,8 @@ const style = {
     'text-align': 'center',
 };
 
-const h3_style = {
-    'font-family': '"Microsoft JhengHei"',
-    'font-size': '36px',
-    'font-weight': 'normal',
-    'line-height': '41px',
-    'letter-spacing': '2px',
-    'color': '#000000',
-};
-
 const Desktop = ({ children }) => (
     <div style={style}>
-        <h3 style={h3_style}>勞動小教室</h3>
         {children}
     </div>
 );
