+++ src/containers/LaborLecture/Lecture.js
@@ -1,0 +1,9 @@
+import { connect } from 'react-redux'
+
+import Lecture from '../../components/LaborLecture/Lecture'
+
+const mapStateToProps = (state, { params: { lecture } }) => ({
+    lecture,
+});
+
+export default connect(mapStateToProps)(Lecture);
