+++ src/routes.js
@@ -2,6 +2,7 @@ import React from 'react';
 import { Route, IndexRoute } from 'react-router';
 import App from './containers/Layout';
 import LaborLecture from './containers/LaborLecture';
+import Lecture from './containers/LaborLecture/Lecture';
 import * as containers from './containers';
 
 
@@ -17,8 +18,11 @@ const routes = () => (
     component={App}
   >
     <IndexRoute component={LandingPage} />
-    <Route path="/labor-lecture" component={LaborLecture} />
-    <Route path="/another" component={AnotherPage} />
+    <Route path="labor-lecture" component={({children}) => children}>
+        <IndexRoute component={LaborLecture}/>
+        <Route path=":lecture" component={Lecture} />
+    </Route>
+    <Route path="another" component={AnotherPage} />
     <Route path="*" component={NotFoundPage} />
   </Route>
 );
